Forked. First commit.
Generalized the REGEXes. Added getopt system to optionally modify the output voltage levels and change the name of the module whose stimulus is rendered as PWL sources. It now works pretty well with Icarus Verilog output.

In addition, the source names reflect the name of the net they drive. Output is word-wrapped at 80 characters.

TODO: The vectored pin code has not been tested with the changes. This is a coming attraction.