Add files via upload
Generalized the REGEX in the script to work with Icarus Verilog. Since what I did was turn hard whitespaces into r"\s+" and r"\s*" constructs, it will probably work with VCDs from other simulators as well.

In addition, I added a getopt routine near the top to provide options to pass the voltage levels you want to use for logic high and low so it won't have to be 0 and 1. The DUT's cell name can also be passed as an option. The original script used "dut", but I set the default to "DUT" to work with my style and, of course, the new command-line option allows a quick change (or you can edit the script!).

Thus far, I only tested it with a block that had only scalar ports. The script had routines to handle buses, but I never had a chance to try them out. It seems likely there are hard-coded whitespaces in there that will need fixing as well.